raftstore: begin extracting fsm::apply
