+++ components/raftstore/src/store/fsm/mod.rs
@@ -1,3 +1,4 @@
+pub mod apply;
 pub mod batch;
 pub mod metrics;
 pub mod router;
